requirements.txt
asgiref==3.8.1
boto3==1.36.5
botocore==1.36.5
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==4.2.18
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
reportlab==4.2.5
requests==2.32.3
s3transfer==0.11.2
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
urllib3==1.26.20
whitenoise==6.8.2
